Tested reading matrix data. Updated cusparse routines to use double.

+++ SpMV_Fortran_OACC/tests/test_cusparse.f90
@@ -5,19 +5,19 @@ program test_cusparse
     integer, parameter :: nrows=4, ncols=4, nnz=9
     integer, dimension(nrows+1):: csrOffsets 
     integer, dimension(nnz) :: columns 
-    real, dimension(nnz) :: values
-    real, dimension(ncols) :: x
-    real, dimension(ncols) :: y, y_result
-    real :: alpha=1.0, beta=0.0
+    real(8), dimension(nnz) :: values
+    real(8), dimension(ncols) :: x
+    real(8), dimension(ncols) :: y, y_result
+    real(8) :: alpha=1.0, beta=0.0
     real :: time_setup=0.0, time_spmv=0.0
 
     csrOffsets = [ 0, 3, 4, 7, 9 ] + 1 ! Change to 1-indexing
     columns    = [ 0, 2, 3, 1, 0, 2, 3, 1, 3 ] + 1 ! Change to 1-indexing
-    values     = [ 1.0, 2.0, 3.0, 4.0, 5.0, &
-                   6.0, 7.0, 8.0, 9.0 ]
-    X        = [  1.0, 2.0,  3.0,  4.0 ]
-    Y        = [  0.0, 0.0,  0.0,  0.0 ]
-    Y_result = [ 19.0, 8.0, 51.0, 52.0 ]
+    values     = [ 1.0d0, 2.0d0, 3.0d0, 4.0d0, 5.0d0, &
+                   6.0d0, 7.0d0, 8.0d0, 9.0d0 ]
+    X        = [  1.0d0, 2.0d0,  3.0d0,  4.0d0 ]
+    Y        = [  0.0d0, 0.0d0,  0.0d0,  0.0d0 ]
+    Y_result = [ 19.0d0, 8.0d0, 51.0d0, 52.0d0 ]
 
     call cusparse_spmv(nrows, ncols, nnz, csrOffsets, columns, values, &
                        x, y, alpha, beta, time_setup, time_spmv)
